requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pinecone-client==2.2.4
openai==1.10.0
pydantic==2.5.3
